refactor: component subfolders
