+++ src/components/forms/NewBandForm.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script>
-    import settings from '../services/settings.js';
+    import settings from '../../services/settings.js';
 
     export default {
         data() {
