+++ src/components/forms/NewPosterForm.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script>
-    import settings from '../services/settings.js';
+    import settings from '../../services/settings.js';
 
     export default {
         data() {
