+++ src/components/mapbox/MapBox.vue
@@ -29,7 +29,7 @@
     import 'mapbox-gl/dist/mapbox-gl.css';
     import { MapboxMap, MapboxMarker } from '@studiometa/vue-mapbox-gl';
     import MapBoxPopup from './MapBoxPopup.vue';
-    import api from '../services/api.js';
+    import api from '../../services/api.js';
 
     export default {
         data() {
