+++ src/components/mapbox/MapBoxDetail.vue
