+++ src/views/Home.vue
@@ -7,7 +7,7 @@
 
 <script>
     import CardContainer from "../components/CardContainer.vue";
-    import MapBox from "../components/MapBox.vue";
+    import MapBox from "../components/mapbox/MapBox.vue";
     
     export default {     
         components: {
